lobster-online-report: migrate existing system test to new infra (#236)
Migrate the valid scenario from the existing system test of the
lobster-online-report to the new infrastructure.
