+++ lobster/tools/core/online_report/requirements.trlc
@@ -15,10 +15,17 @@ req.Software_Requirement Dummy_Requirement_Unit_Test {
   '''
 }
 
+req.System_Requirement Valid_Git_Repository {
+  description = '''
+    IF the provided repository root directory contains a ".git" directory,
+    THEN the tool shall write the output in the LOBSTER interchange format to the file given in the "out" parameter
+    AND exit with a zero return code.
+  '''
+}
+
 req.System_Requirement_Aspect Invalid_Git_Repository {
   description = '''
-    IF the provided repository root directory does not contain a ".git" directory,
-    THEN the tool shall exit with a non-zero return code and display an error message
+    OTHERWISE the tool shall exit with a non-zero return code and display an error message
     indicating that the repository is not a valid Git repository.
   '''
 }
