+++ tests-system/lobster-online-report/lobster_online_report_system_test_case_base.py
@@ -10,8 +10,10 @@ class LobsterOnlineReportSystemTestCaseBase(SystemTestCaseBase):
 
     def create_test_runner(self) -> LobsterOnlineReportTestRunner:
         tool_name = Path(__file__).parents[0].name
+        working_dir = Path(__file__).parents[2]
         test_runner = LobsterOnlineReportTestRunner(
             tool_name,
-            self.create_temp_dir(prefix=f"test-{tool_name}-"),
+            self.create_temp_dir(prefix=f"test-{tool_name}-",
+                                 dir_path=working_dir),
         )
         return test_runner
