--- tests-system/lobster-online-report/rbt-valid-flow/valid-scenario/expected-output/exit-code.txt
@@ -1,1 +1,0 @@
-0
