--- tests-system/lobster-online-report/rbt-valid-flow/valid-scenario/expected-output/stdout.txt
@@ -1,1 +1,0 @@
-LOBSTER report expected-output.lobster created, using online references.
