--- tests-system/lobster-online-report/rbt-valid-flow/valid-scenario/input/args.txt
@@ -1,2 +1,0 @@
-report.lobster
---out=expected-output.lobster
